player_simple_working_directory_v7/requirements.txt
pyserial>=3.5
pytest>=7.0.0
pillow
